refactor: Move new integration test
